Add casts to shut up warnings

+++ src/leftside_logic.c
@@ -9,8 +9,8 @@ static int deploy_cooldown = 0;
 static void reroll_item(int i)
 {
     items_conveyor.shapes[i] = GetRandomShape();
-    items_conveyor.positions[i].x = GetRandomValue(0, LS_LOGICAL_WIDTH - items_conveyor.shapes[i]->art.width);
-    items_conveyor.positions[i].y = -items_conveyor.shapes[i]->art.height - deploy_cooldown * conveyor_velocity;
+    items_conveyor.positions[i].x = (float)GetRandomValue(0, LS_LOGICAL_WIDTH - items_conveyor.shapes[i]->art.width);
+    items_conveyor.positions[i].y = (float)-items_conveyor.shapes[i]->art.height - deploy_cooldown * conveyor_velocity;
     deploy_cooldown += items_conveyor.shapes[i]->art.height / conveyor_velocity;
     items_conveyor.active[i] = 1;
 }
@@ -58,8 +58,8 @@ void leftside_logic()
         //mouse in bounding box
         if (items_conveyor.active[RING_INDEX_RAW(items_conveyor, i)] && IsMouseButtonPressed(MOUSE_LEFT_BUTTON) && GetMouseX() >= RING_INDEX_POS(items_conveyor, i).x && GetMouseX() <= RING_INDEX_POS(items_conveyor, i).x + RING_INDEX_IDS(items_conveyor, i)->art.width && GetMouseY() >= RING_INDEX_POS(items_conveyor, i).y && GetMouseY() <= RING_INDEX_POS(items_conveyor, i).y + RING_INDEX_IDS(items_conveyor, i)->art.height)
         {
-            mouse_block_x = (GetMouseX() - RING_INDEX_POS(items_conveyor, i).x) / (RING_INDEX_IDS(items_conveyor, i)->art.width / 8);
-            mouse_block_y = (GetMouseY() - RING_INDEX_POS(items_conveyor, i).y) / (RING_INDEX_IDS(items_conveyor, i)->art.height / 8);
+            mouse_block_x = (int)(GetMouseX() - RING_INDEX_POS(items_conveyor, i).x) / (RING_INDEX_IDS(items_conveyor, i)->art.width / 8);
+            mouse_block_y = (int)(GetMouseY() - RING_INDEX_POS(items_conveyor, i).y) / (RING_INDEX_IDS(items_conveyor, i)->art.height / 8);
             if (RING_INDEX_IDS(items_conveyor, i)->grid[mouse_block_y][mouse_block_x] == 1)
             {
                 current_hold_item.shape = RING_INDEX_IDS(items_conveyor, i);
